Add files via upload
Hey Abdul, this is my proposal mostly regarding the structure. In section 1 and 2 I generated many plots: we will choose together with the others which one are more communicative and straightforward. I am struggling in the 3rd section. Have a look at it, to see what I was trying to do. For sure there are many things that have to be fixed but let me know if you think the structure is reasonable.